// Gdc.Scd.Web.Client/src/CapabilityMatrix/Model/CapabilityMatrixFilterModel.ts
export interface CapabilityMatrixFilterModel {
    country?: string;
    wg?: string;
    availability?: string;
    duration?: string;
    reactionType?: string;
    reactionTime?: string;
    serviceLocation?: string;
    proActive?: string;

    isGlobalPortfolio?: boolean;
    isMasterPortfolio?: boolean;
    isCorePortfolio?: boolean;
}